Enemy behaviors added
Multiple new movement patterns added with prefabs. The movement code was also refactored to make things easier. Also added new graphics for the spinning multishot enemy type. I'll have to create a set of enemies that are the most interesting in game but for now the options are good for experimentation. Next up is melee attacks that currently have a little to no implementation.